set Authoring react component as unique soruce of truth for sidebar state
