+++ assets/src/apps/authoring/BottomPanel.tsx
@@ -19,11 +19,11 @@ export interface BottomPanelProps {
   onToggle: any;
   children?: any;
   content?: any;
-  initialSidebarExpanded?: boolean;
+  sidebarExpanded?: boolean;
 }
 
 export const BottomPanel: React.FC<BottomPanelProps> = (props: BottomPanelProps) => {
-  const { panelState, onToggle, children, initialSidebarExpanded } = props;
+  const { panelState, onToggle, children, sidebarExpanded } = props;
   const PANEL_SIDE_WIDTH = '270px';
   const dispatch = useDispatch();
   const currentRule = useSelector(selectCurrentRule);
@@ -34,11 +34,6 @@ export const BottomPanel: React.FC<BottomPanelProps> = (props: BottomPanelProps)
   const isLayer = getIsLayer();
 
   const ref = useRef<HTMLElement>(null);
-  const [sidebarExpanded, setSidebarExpanded] = useState(initialSidebarExpanded);
-
-  const handleSidebarExpanded = () => {
-    setSidebarExpanded((prev) => !prev);
-  };
 
   useEffect(() => {
     if (currentRule === undefined) return;
@@ -160,13 +155,6 @@ export const BottomPanel: React.FC<BottomPanelProps> = (props: BottomPanelProps)
 
   return (
     <>
-      <button
-        role="update sidebar state on React"
-        className="hidden"
-        onClick={() => {
-          handleSidebarExpanded();
-        }}
-      ></button>
       <section
         id="aa-bottom-panel"
         ref={ref}
