+++ assets/src/apps/authoring/components/ExpertHeaderNav.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useState } from 'react';
 import { OverlayTrigger, Tooltip } from 'react-bootstrap';
 import { useDispatch, useSelector } from 'react-redux';
 import {
@@ -21,11 +20,11 @@ interface HeaderNavProps {
   isVisible: boolean;
   authoringContainer: React.RefObject<HTMLElement>;
   onToggleExport?: () => void;
-  initialSidebarExpanded?: boolean;
+  sidebarExpanded?: boolean;
 }
 
 const ExpertHeaderNav: React.FC<HeaderNavProps> = (props: HeaderNavProps) => {
-  const { panelState, isVisible, initialSidebarExpanded } = props;
+  const { panelState, isVisible, sidebarExpanded } = props;
   const projectSlug = useSelector(selectProjectSlug);
   const revisionSlug = useSelector(selectRevisionSlug);
   const paths = useSelector(selectPaths);
@@ -37,12 +36,6 @@ const ExpertHeaderNav: React.FC<HeaderNavProps> = (props: HeaderNavProps) => {
 
   const url = `/authoring/project/${projectSlug}/preview/${revisionSlug}`;
   const windowName = `preview-${projectSlug}`;
-
-  const [sidebarExpanded, setSidebarExpanded] = useState(initialSidebarExpanded);
-
-  const handleSidebarExpanded = () => {
-    setSidebarExpanded((prev) => !prev);
-  };
 
   const handleReadOnlyClick = () => {
     // TODO: show a modal offering to confirm if you want to disable read only
@@ -69,13 +62,6 @@ const ExpertHeaderNav: React.FC<HeaderNavProps> = (props: HeaderNavProps) => {
           right: panelState['right'] ? PANEL_SIDE_WIDTH : 0,
         }}
       >
-        <button
-          role="update sidebar state on React"
-          className="hidden"
-          onClick={() => {
-            handleSidebarExpanded();
-          }}
-        ></button>
         <div className="btn-toolbar" role="toolbar">
           <div className="btn-group pl-3 align-items-center" role="group" aria-label="Third group">
             <UndoRedoToolbar />
